Fix NPE issue
Some corrections are made to the plugin.xml so that the Queue
resources can be correctly queried and retrieved and shown
on the console.
Also we need a way to resolve "brokerName" from grandchildren
resources. Currently only direct children to the root broker
can have this value provided. For grandchildren we setup
a static Map to store the "brokerName" for them to retrieve.