AMQ role based access control support
AMQ Jon plugin supports AMQ 7's RBAC management feature. To be able to
have the right access to broker's management resource, Jon must
use a valid credential (username/password) to connect to the
broker. The broker uses JAAS login module to authenticate
against Jon's credentials and then authorizes any requests from
Jon plugin against the access-control policy defined in
management.xml.

The Jon plugin's default user is 'admin' and default password is 'activemq'.
You can configure the plugin to use a different user/password to connect
to the broker. The configuration file are located in the following path:

<artemis.instance>/etc/org.jboss.rh-messaging.amq.jon.cfg

This file takes the form of java properties file. It allows you to
specify the user name and password for the Jon plugin to connect to
the broker. You can also configure the jmx connector url in this file.

For example a configuration can have the following properties:

principal = guest
credentials = guest
connectorAddress = service:jmx:rmi:///jndi/rmi://localhost:11099/jmxrmi

where the value of 'principal' is the user name (i.e. guest); the
value of 'credentials' is password (i.e. guest); and the value
of 'connectorAddress' is the connector's url.